m9u.go: add playlist index to /event
Changes the form of the event messages:

 Play <n> [P|Q|C] <song_path>
 Stop <n> [P|Q] <song_path>
 Stop !

The first field (Play or Stop) indicates whether playback is currently
active.

The second field gives the current playlist position. This is a 1-based
index into /list.

The third field indicates the source of the path:
  P => <song_path> came from the playlist
  Q => <song_path> came from the queue
  C => <song_path> came from the ctl file

The rest of the line contains the path to the song.


Some notes on interpreting event messages:

* <song_path> in a Play message refers to the currently playing song
* <song_path> in a Stop message refers to the song next to be played,
    which may come from either the playlist or queue (see 3rd field)
* when the 3rd field reads "P", <song_path> will match the <n>-th
    song in /list (unless /list has been modified since the event
    message was generated)
* when the 3rd field is not "P", <n> is not immediately useful but
    tells you which song in the playlist will be played once the
    current song finishes and the queue is emptied


This commit also changes /queue semantics - a song now remains on top
of the queue until it plays through or is explicitly skipped
(previously the song was removed from the queue immediately after it
began playing).

This additionally affects the behaviour of the 'skip' ctl message;
previously it always modified the playlist position but now if something
is on the queue the argument to skip is ignored and a single song from
the top of queue removed.